backend/src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Core
    "fastapi>=0.115.13",
    "uvicorn>=0.34.3",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",

    # Auth, config
    "pydantic==2.9.1",                     
    "pydantic-settings>=2.9.1",
    "python-jose>=3.5.0",
    "python-dotenv[cli]~=1.0.0",

    # Networking
    "httpx~=0.27.2",                        
    "requests~=2.32.3",
    "aiohttp>=3.10.11",
    "anyio==4.4.0",
    "urllib3==2.2.3",
    "certifi==2024.8.30",
    "idna==3.7",

    # Langchain ecosystem
    "langchain~=0.3.24",
    "langchain-core~=0.3.35",
    "langchain-community==0.3.17",
    "langchain-openai~=0.3.14",
    "langchain-google-genai==2.0.6",
    "langchain-cohere==0.3.3",
    "langchain-together==0.3.0",
    "langchain-ollama==0.2.1",
    "langchain_deepseek==0.1.3",

    # Graph-based langchain (đã có sẵn)
    "langgraph>=0.3.24,<0.4.0",

    # AI tools
    "ollama~=0.4.7",
    "huggingface-hub~=0.24.5",
    "tavily-python>=0.5.4",

    # PDF & file parsing
    "pdf2image",
    "pypdf~=5.4.0",
    "lxml~=4.9.3",
    "python-pptx~=1.0.2",
    "json5~=0.9.14",
    "json-repair==0.29.2",

    # Visualization & UI
    "streamlit~=1.44.1",
    "matplotlib~=3.10.1",
    "ipython~=8.36.0",

    # NLP/Data
    "numpy",
    "scikit-learn~=1.5.1",
    "pyarrow~=16.0.0",

    # Utils
    "jinja2>=3.1.5",
    "Pillow==10.3.0",
    "tqdm~=4.66.5",
    "gitpython==3.1.43"
]
